be a little safer in how we kill our ipcclient worker

+++ Dev/Warewolf.COMIPC.Client/IpcClient.cs
@@ -90,7 +90,17 @@ namespace WarewolfCOMIPC.Client
         protected void Close()
         {
             _pipeWrapper.Close();
-            _process?.Kill();
+
+            while (!_process.HasExited)
+            {
+                if (!_process.WaitForExit(1000))
+                {
+                    try
+                    {
+                        _process?.Kill();
+                    } catch {}
+                }
+            }
         }
 
         ~IpcClientImpl()
